Move operating table module to ui
